refac: directoring game simulation function
